♻️  Extract 3 functions to clarify the query
